+++ battleship.rb
@@ -3,3 +3,6 @@ require './player'
 require './ship'
 require './grid'
 require './position'
+require './game'
+require './computer'
+require './humanplayer'
