+++ game.rb
@@ -1,0 +1,2 @@
+class Game
+end
